Important changes
-The course.js and auth.js were deleted because we didnt used and Dash suggest to erase them
-Chages in the index.js,in the importation routes
-The verify method of the authController.js was moved to the is-auth.js in the middlewares  carpet
-The  previous was executed beacause the verifyToken act  as a middleware in some requests
-The config/config.js was modified adding the environment variables
-The routes/searcher.js was modified  because verifytoken now is on the middlewares carpet
