+++ config/config.js
@@ -1,8 +1,8 @@
 module.exports = {
   development: {
     username: 'root',
-    password: '',
-    database: '',
+    password: process.env.DB_PASSWORD,
+    database: process.env.DB_NAME,
     host: '127.0.0.1',
     port: 3306,
     dialect: 'mysql',
