+++ controllers/authController.js
@@ -61,17 +61,3 @@ exports.login = async (req, res, next) => {
   }
 };
 
-exports.verifyToken=(req,res,next)=>{
-  let token=req.get('token');//Authenticate
-  jwt.verify(token,process.env.JWT_SECRET,(error,decoded)=>{
-    //decoded is the payload
-    if(error){
-      return res.status(401).json({
-        ok:false,
-        error
-      });
-    }
-    req.user=decoded.user;  
-  });
-  next();
-}
