+++ routes/searcher.js
@@ -1,10 +1,10 @@
 const express = require("express");
 const router = express.Router();
-const is_auth = require("../middlewares/is-auth");
+const {verifyToken} = require("../middlewares/is-auth");
 const searchController = require("../controllers/searchController");
 const searchCategoryController = require("../controllers/searcherCategoryController");
 
-router.get("/search", is_auth, searchController.searcher);
-router.get("/search/category/:category", is_auth, searchCategoryController.searcher)
+router.get("/search", verifyToken, searchController.searcher);
+router.get("/search/category/:category", verifyToken, searchCategoryController.searcher)
 
 module.exports = router;
